Handle API update to better handle OAuth.
There was a build break because the go-autorest changed the way Service
Principal Tokens (SPT) are created.  The API is now explicity about how an
OAuth token is formulated, and the endpoints used.  The code hardcodes to
the public cloud.  We are no worse than we were because we still did not
support US Government Cloud and China.  There is already a pending issue
to address this.