Support interrupting long running tasks.
Fix for #141.

If a user hits Ctrl-C the user will have to wait until the step has
completed before the Ctrl-C is processed.  This change processes the
Ctrl-C within a 100ms, and immediately aborts all work.  There is no
cleanup done, and the task running in Azure is not killed.  The following
tasks are now interruptible.

 * capture image
 * delete resource group
 * deploy template
 * power off

All othere tasks run quickly enough that I did not make them
interruptible.

Small refactoring to the way errors are handled at the end of each step.