Log Azure requests and responses.
The body is currently captured because it contains valuable information.
The downside is that the body is very big, and makes very noisy logs.  I'm
not sure if the happy medium is not to log the body because without it
diagnosing failures is very difficult.  The environment variable
PACKER_LOG_AZURE_MAXLEN can be used to limited the log size of request and
response bodies.